[refactor]: restructure agents system architecture
