+++ backend/app/agents/__init__.py
@@ -1,6 +1,6 @@
 from .devrel.agent import DevRelAgent
-from .shared.base_agent import BaseAgent, AgentState
-from .shared.classification_router import ClassificationRouter
+from .base_agent import BaseAgent, AgentState
+from .classification_router import ClassificationRouter
 
 __all__ = [
     "DevRelAgent",
