+++ backend/app/agents/devrel/nodes/handlers/web_search.py
@@ -1,12 +1,15 @@
 import logging
-from app.agents.shared.state import AgentState
+from typing import Dict, Any
+from app.agents.state import AgentState
 from langchain_core.messages import HumanMessage
-from ..prompts.search_prompt import EXTRACT_SEARCH_QUERY_PROMPT
+from ...prompts.search_prompt import EXTRACT_SEARCH_QUERY_PROMPT
 
 logger = logging.getLogger(__name__)
 
 async def _extract_search_query(message: str, llm) -> str:
-    """Extract a concise search query from the user's message."""
+    """
+    Extract a concise search query from the user's message by invoking the LLM.
+    """
     logger.info(f"Extracting search query from: {message[:100]}")
     try:
         prompt = EXTRACT_SEARCH_QUERY_PROMPT.format(message=message)
@@ -19,7 +22,9 @@ async def _extract_search_query(message: str, llm) -> str:
     return search_query
 
 async def handle_web_search_node(state: AgentState, search_tool, llm) -> dict:
-    """Handle web search requests"""
+    """
+    Handle web search requests
+    """
     logger.info(f"Handling web search for session {state.session_id}")
 
     latest_message = ""
@@ -41,3 +46,25 @@ async def handle_web_search_node(state: AgentState, search_tool, llm) -> dict:
         "tools_used": ["tavily_search"],
         "current_task": "web_search_handled"
     }
+
+def create_search_response(task_result: Dict[str, Any]) -> str:
+    """
+    Create a user-friendly response string from search results.
+    """
+    query = task_result.get("query")
+    results = task_result.get("results", [])
+
+    if not results:
+        return f"I couldn't find any information for '{query}'. You might want to try rephrasing your search."
+
+    response_parts = [f"Here's what I found for '{query}':"]
+    for i, result in enumerate(results[:5]):
+        title = result.get('title', 'N/A')
+        snippet = result.get('snippet', 'N/A')
+        url = result.get('url', '#')
+        result_line = f"{i+1}. {title}: {snippet}"
+        response_parts.append(result_line)
+        response_parts.append(f"   (Source: {url})")
+
+    response_parts.append("You can ask me to search again with a different query if these aren't helpful.")
+    return "\n".join(response_parts)
